+++ lib/faraday/detailed_logger.rb
@@ -1,7 +1,12 @@
 # frozen_string_literal: true
 
-require 'faraday/detailed_logger/version'
-require 'faraday/detailed_logger/middleware'
+require 'zeitwerk'
+loader = Zeitwerk::Loader.new
+loader.tag = 'faraday-detailed_logger'
+loader.inflector = Zeitwerk::GemInflector.new(__FILE__)
+loader.push_dir(File.join(__dir__, '..'))
+loader.ignore(__FILE__)
+loader.setup
 
 module Faraday
   module DetailedLogger