+++ lib/faraday/detailed_logger/middleware.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require 'faraday'
-require 'faraday/detailed_logger/tagged_logging'
 
 module Faraday
   module DetailedLogger